+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## Version 0.2.0 (not yet)
 
+- add the "-x filename" argument to netplan-parser
+
 ## Version 0.1.0 (2018-07-31)
 
 - initial public release